fix turret building menu
